+++ ohsome_quality_api/geodatabase/client.py
@@ -17,6 +17,7 @@ import os
 from contextlib import asynccontextmanager
 
 import asyncpg
+import geojson
 from asyncpg import Record
 from geojson import Feature, FeatureCollection
 
@@ -98,3 +99,15 @@ async def get_eubucco_coverage_intersection_area(bpoly: Feature) -> list[Record]
     geom = str(bpoly.geometry)
     async with get_connection() as conn:
         return await conn.fetch(query, geom)
+
+
+async def get_eubucco_coverage_intersection(bpoly: Feature) -> Feature:
+    """Get intersection geometry of AoI and coverage geometry."""
+    file_path = os.path.join(WORKING_DIR, "get_coverage_intersection.sql")
+    with open(file_path, "r") as file:
+        query = file.read()
+    geom = str(bpoly.geometry)
+    async with get_connection() as conn:
+        result = await conn.fetch(query, geom)
+        bpoly["geometry"] = geojson.loads(result[0]["geom"])
+        return bpoly