+++ tests/integrationtests/indicators/test_building_comparison.py
@@ -14,7 +14,16 @@ from tests.integrationtests.utils import oqapi_vcr
 
 @pytest.fixture(scope="class")
 def mock_get_building_area(class_mocker):
-    async_mock = AsyncMock(return_value=[{"area": 4842587.791645115 / (1000 * 1000)}])
+    async_mock = AsyncMock(return_value=[{"area": 4842587.791645115}])
+    class_mocker.patch(
+        "ohsome_quality_api.indicators.building_completeness.indicator.db_client.get_building_area",
+        side_effect=async_mock,
+    )
+
+
+@pytest.fixture(scope="class")
+def mock_get_building_area_low(class_mocker):
+    async_mock = AsyncMock(return_value=[{"area": 1}])
     class_mocker.patch(
         "ohsome_quality_api.indicators.building_completeness.indicator.db_client.get_building_area",
         side_effect=async_mock,
@@ -23,9 +32,27 @@ def mock_get_building_area(class_mocker):
 
 @pytest.fixture(scope="class")
 def mock_get_building_area_empty(class_mocker):
-    async_mock = AsyncMock(return_value=[])
+    async_mock = AsyncMock(return_value=[{"area": 0}])
     class_mocker.patch(
         "ohsome_quality_api.indicators.building_completeness.indicator.db_client.get_building_area",
+        side_effect=async_mock,
+    )
+
+
+@pytest.fixture(scope="class")
+def mock_get_eubucco_coverage_intersection(class_mocker, feature_germany_berlin):
+    async_mock = AsyncMock(return_value=feature_germany_berlin)
+    class_mocker.patch(
+        "ohsome_quality_api.indicators.building_completeness.indicator.db_client.get_eubucco_coverage_intersection",
+        side_effect=async_mock,
+    )
+
+
+@pytest.fixture(scope="class")
+def mock_get_eubucco_coverage_intersection_area(class_mocker):
+    async_mock = AsyncMock(return_value=[{"area_ratio": 1}])
+    class_mocker.patch(
+        "ohsome_quality_api.indicators.building_completeness.indicator.db_client.get_eubucco_coverage_intersection_area",
         side_effect=async_mock,
     )
 
@@ -45,6 +72,8 @@ class TestPreprocess:
         mock_get_building_area,
         topic_building_area,
         feature_germany_berlin,
+        mock_get_eubucco_coverage_intersection_area,
+        mock_get_eubucco_coverage_intersection,
     ):
         indicator = BuildingComparison(topic_building_area, feature_germany_berlin)
         asyncio.run(indicator.preprocess())
@@ -52,7 +81,7 @@ class TestPreprocess:
         assert indicator.area_osm is not None
         assert indicator.area_osm > 0
         assert indicator.area_references == {
-            "EUBUCCO": 4.842587791645116 / (1000 * 1000)
+            "EUBUCCO": 4.842587791645116,
         }
         assert isinstance(indicator.result.timestamp, datetime)
         assert isinstance(indicator.result.timestamp_osm, datetime)
@@ -65,6 +94,8 @@ class TestCalculate:
         mock_get_building_area,
         topic_building_area,
         feature_germany_berlin,
+        mock_get_eubucco_coverage_intersection_area,
+        mock_get_eubucco_coverage_intersection,
     ):
         indicator = BuildingComparison(topic_building_area, feature_germany_berlin)
         asyncio.run(indicator.preprocess())
@@ -80,12 +111,37 @@ class TestCalculate:
         mock_get_building_area_empty,
         topic_building_area,
         feature_germany_heidelberg,
+        mock_get_eubucco_coverage_intersection_area,
+        mock_get_eubucco_coverage_intersection,
     ):
         indicator = BuildingComparison(topic_building_area, feature_germany_heidelberg)
         asyncio.run(indicator.preprocess())
         indicator.coverage["EUBUCCO"] is None
         indicator.calculate()
         assert indicator.result.value is None
+
+    @oqapi_vcr.use_cassette
+    def test_calculate_above_one_th(
+        self,
+        mock_get_building_area_low,
+        topic_building_area,
+        feature_germany_heidelberg,
+        mock_get_eubucco_coverage_intersection_area,
+        mock_get_eubucco_coverage_intersection,
+    ):
+        indicator = BuildingComparison(topic_building_area, feature_germany_heidelberg)
+        asyncio.run(indicator.preprocess())
+        indicator.calculate()
+        assert indicator.result.value is not None
+        assert indicator.result.value > 0
+        assert indicator.result.class_ is None
+        assert indicator.result.description is not None
+        assert (
+            "Warning: No quality estimation made. "
+            "OSM and reference data differ. Reference data is likely outdated."
+            in indicator.result.description
+        )
+        assert indicator.result.label == "undefined"
 
 
 class TestFigure:
@@ -96,6 +152,8 @@ class TestFigure:
         mock_get_building_area,
         topic_building_area,
         feature_germany_berlin,
+        mock_get_eubucco_coverage_intersection_area,
+        mock_get_eubucco_coverage_intersection,
     ):
         indicator = BuildingComparison(topic_building_area, feature_germany_berlin)
         asyncio.run(indicator.preprocess())
@@ -111,3 +169,37 @@ class TestFigure:
     def test_create_figure_manual(self, indicator):
         indicator.create_figure()
         pio.show(indicator.result.figure)
+
+    @oqapi_vcr.use_cassette
+    def test_create_figure_above_one_th(
+        self,
+        mock_get_building_area_low,
+        topic_building_area,
+        feature_germany_berlin,
+        mock_get_eubucco_coverage_intersection_area,
+        mock_get_eubucco_coverage_intersection,
+    ):
+        indicator = BuildingComparison(topic_building_area, feature_germany_berlin)
+        asyncio.run(indicator.preprocess())
+        indicator.calculate()
+        indicator.create_figure()
+        assert isinstance(indicator.result.figure, dict)
+        assert indicator.result.figure["data"][0]["type"] == "bar"
+        pgo.Figure(indicator.result.figure)
+
+    @oqapi_vcr.use_cassette
+    def test_create_figure_building_area_zero(
+        self,
+        mock_get_building_area_empty,
+        topic_building_area,
+        feature_germany_berlin,
+        mock_get_eubucco_coverage_intersection_area,
+        mock_get_eubucco_coverage_intersection,
+    ):
+        indicator = BuildingComparison(topic_building_area, feature_germany_berlin)
+        asyncio.run(indicator.preprocess())
+        indicator.calculate()
+        indicator.create_figure()
+        assert isinstance(indicator.result.figure, dict)
+        assert indicator.result.figure["data"][0]["type"] == "bar"
+        pgo.Figure(indicator.result.figure)