+++ tests/unittests/test_indicators_definitions.py
@@ -1,8 +1,21 @@
 import asyncio
+from unittest.mock import AsyncMock
 
 import geojson
+import pytest
 
 from ohsome_quality_api.indicators import definitions, models
+
+
+@pytest.fixture(scope="class")
+def mock_select_eubucco_coverage(class_mocker, feature_germany_berlin):
+    async_mock = AsyncMock(
+        return_value=[{"geom": geojson.dumps(feature_germany_berlin)}]
+    )
+    class_mocker.patch(
+        "ohsome_quality_api.indicators.building_completeness.indicator.db_client.get_eubucco_coverage",
+        side_effect=async_mock,
+    )
 
 
 def test_get_indicator_names():
@@ -30,7 +43,7 @@ def test_get_indicator_definitions_with_project():
         assert indicator.projects == ["core"]
 
 
-def test_get_coverage():
+def test_get_coverage(mock_select_eubucco_coverage):
     coverage = asyncio.run(definitions.get_coverage("building-comparison"))
     assert coverage.is_valid
     assert isinstance(coverage, geojson.FeatureCollection)